Switched doubles to long double to describe position in the complex plane so we can zoom in more

+++ sources/map.c
@@ -6,12 +6,12 @@
 /*   By: lucas <lopoka@student.hive.fi>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/04 15:31:10 by lucas             #+#    #+#             */
-/*   Updated: 2024/06/04 15:31:30 by lucas            ###   ########.fr       */
+/*   Updated: 2024/06/05 16:36:05 by lopoka           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "../includes/fractol.h" 
 
-double	ft_map(double val, t_fract *stc, int axis)
+long double	ft_map(long double val, t_fract *stc, int axis)
 {
 	if (!axis)
 		return (val * (stc->max_y - stc->min_y) / HEIGHT + stc->min_y);
